pyproject.toml
[project]
name = "product_description_generator"
version = "0.1.0"
description = "Product Description Generator using CrewAI"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.76.2",
    "langchain-core>=0.2.30",
    "langchain-google-genai>=0.0.6",
    "google-cloud-storage>=2.10.0",
]

[project.scripts]
product_description_generator = "product_description_generator.main:run"
run_crew = "product_description_generator.main:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
